Refactor utils funcs
