+++ functions/package.json
@@ -30,6 +30,7 @@
     "@babel/core": "^7.18.2",
     "@babel/preset-env": "^7.18.2",
     "@babel/preset-typescript": "^7.17.12",
+    "@types/sinon": "^10.0.13",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
     "babel-jest": "^28.1.1",
@@ -42,6 +43,7 @@
     "firebase-functions-test": "^0.2.3",
     "jest-mock-axios": "^4.6.1",
     "mocha": "^10.0.0",
+    "sinon": "^14.0.0",
     "ts-node": "^10.8.1",
     "typescript": "^4.5.4",
     "wait-for-expect": "^3.0.2"
